% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findInGit.R
\name{FIG2dataframe}
\alias{FIG2dataframe}
\title{Output of `findInGit` as a dataframe}
\usage{
FIG2dataframe(fig)
}
\arguments{
\item{fig}{the output of \code{\link{findInGit}} used with the
option \code{output = "viewer+dataframe"}}
}
\value{
The results of \code{\link{findInGit}} in a dataframe.
}
\description{
Returns the results of \code{\link{findInGit}} in a dataframe, when the
  option \code{output = "viewer+dataframe"} is used. See the example in
  \code{\link{findInGit}}.
}
